+++ example/cumulus-tf/parse_pdr_workflow.tf
@@ -5,7 +5,7 @@ module "parse_pdr_workflow" {
   name            = "ParsePdr"
   workflow_config = module.cumulus.workflow_config
   system_bucket   = var.system_bucket
-  tags            = local.default_tags
+  tags            = local.tags
 
   state_machine_definition = <<JSON
 {