+++ example/cumulus-tf/sync_granule_catch_duplicate_error_test_workflow.tf
@@ -5,7 +5,7 @@ module "sync_granule_catch_duplicate_error_test" {
   name            = "SyncGranuleCatchDuplicateErrorTest"
   workflow_config = module.cumulus.workflow_config
   system_bucket   = var.system_bucket
-  tags            = local.default_tags
+  tags            = local.tags
 
   state_machine_definition = <<JSON
 {